moved document search bar from entity breadcrumsb to action bar
