+++ ui/src/screens/EntityScreen/EntityScreen.jsx
@@ -4,7 +4,6 @@ import { defineMessages, injectIntl } from 'react-intl';
 import queryString from 'query-string';
 import { ButtonGroup } from '@blueprintjs/core';
 
-import Query from 'app/Query';
 import Screen from 'components/Screen/Screen';
 import EntityContextLoader from 'components/Entity/EntityContextLoader';
 import { compose } from 'redux';
@@ -16,19 +15,13 @@ import EntityDeleteButton from 'components/Toolbar/EntityDeleteButton';
 import LoadingScreen from 'components/Screen/LoadingScreen';
 import ErrorScreen from 'components/Screen/ErrorScreen';
 import EntitySetSelector from 'components/EntitySet/EntitySetSelector';
-import { Breadcrumbs, DualPane, SearchBox } from 'components/common';
+import { Breadcrumbs, DualPane } from 'components/common';
 import { DialogToggleButton } from 'components/Toolbar';
 import { DownloadButton } from 'components/Toolbar';
 import { deleteEntity } from 'actions';
-import { queryEntityReference } from 'queries';
-import {
-  selectEntity, selectEntityReference, selectEntityView,
-} from 'selectors';
+import { selectEntity, selectEntityView } from 'selectors';
 
 import 'components/common/ItemOverview.scss';
-
-const SEARCHABLES = ['Pages', 'Folder', 'Package', 'Workbook'];
-
 
 const messages = defineMessages({
   add_to: {
@@ -38,29 +31,9 @@ const messages = defineMessages({
 });
 
 class EntityScreen extends Component {
-  constructor(props) {
-    super(props);
-
-    this.onSearch = this.onSearch.bind(this);
-  }
-
-  onSearch(queryText, entityLink) {
-    const { history, location, query } = this.props;
-    const parsedHash = queryString.parse(location.hash);
-    const newQuery = query.setString('q', queryText);
-    parsedHash['preview:id'] = undefined;
-    parsedHash['preview:mode'] = undefined;
-    parsedHash.page = undefined;
-    history.push({
-      pathname: entityLink,
-      search: newQuery.toLocation(),
-      hash: queryString.stringify(parsedHash),
-    });
-  }
-
   render() {
     const {
-      entity, entityId, activeMode, query, intl,
+      entity, entityId, activeMode, intl,
     } = this.props;
 
     if (entity.isError) {
@@ -75,15 +48,6 @@ class EntityScreen extends Component {
     }
 
     const { writeable } = entity.collection;
-    const hasSearch = entity.schema.isAny(SEARCHABLES) && !entity.schema.isA('Email') && !entity.schema.isA('Folder');
-
-    const search = hasSearch && (
-      <SearchBox
-        onSearch={this.onSearch}
-        placeholderLabel={entity.getCaption()}
-        query={query}
-      />
-    );
 
     const operation = (
       <ButtonGroup>
@@ -113,7 +77,7 @@ class EntityScreen extends Component {
     );
 
     const breadcrumbs = (
-      <Breadcrumbs operation={operation} search={search} >
+      <Breadcrumbs operation={operation} >
         <Breadcrumbs.Collection collection={entity.collection} />
         <Breadcrumbs.Entity entity={entity} />
       </Breadcrumbs>
@@ -147,23 +111,16 @@ class EntityScreen extends Component {
 }
 
 const mapStateToProps = (state, ownProps) => {
-  const { entityId } = ownProps.match.params;
-  const { location } = ownProps;
+  const { location, match } = ownProps;
+  const { entityId } = match.params;
   const entity = selectEntity(state, entityId);
   const hashQuery = queryString.parse(location.hash);
-  const isDocument = entity?.schema?.isDocument();
   const activeMode = selectEntityView(state, entityId, hashQuery.mode, false);
-  const reference = selectEntityReference(state, entityId, activeMode);
-  const referenceQuery = queryEntityReference(location, entity, reference);
-  const documentQuery = Query.fromLocation('entities', location, {}, 'document');
 
   return {
     entity,
     entityId,
-    reference,
     activeMode,
-    isDocument,
-    query: referenceQuery || documentQuery,
   };
 };
